Switch the release baseline from jdk25u to the OpenJDK mainline (JDK 27)
carried on build-jdk27. This commit takes build-jdk27's tree wholesale so
that future mainline updates can be merged cleanly; ByteOpenJDK-specific
customizations (CI pipeline, SweeperThreshold, AVX-512 IceLake guard) are
re-applied in the following commits.

+++ src/java.net.http/share/classes/jdk/internal/net/http/common/TimeSource.java
@@ -25,7 +25,6 @@
 package jdk.internal.net.http.common;
 
 import java.time.Instant;
-import java.time.InstantSource;
 
 /**
  * A {@link TimeLine} based on {@link System#nanoTime()} for the
@@ -52,7 +51,7 @@ public final class TimeSource implements TimeLine {
         // The use of Integer.MAX_VALUE is arbitrary.
         // Any value not too close to Long.MAX_VALUE
         // would do.
-        static final int TIME_WINDOW = Integer.MAX_VALUE;
+        static final long TIME_WINDOW = Integer.MAX_VALUE;
 
         final Instant first;
         final long firstNanos;
